requirements.txt
google-cloud-storage
google-cloud-bigquery
websocket-client
requests
